DCMAW-12496: Move crypto to mobile/utils
